Refactors helper module structure for type-checking
Renames and reorganizes the helper file to improve clarity and align
its purpose with type-checking functionality. Updates import paths
to reflect the new file structure, ensuring consistent module references.
